LPS-109189 Move BundleUtil to a more specific package, com.liferay.portal.module.util
