+++ modules/apps/portal/portal-upgrade-impl/src/main/java/com/liferay/portal/upgrade/internal/index/updater/osgi/commands/IndexUpdaterOSGiCommands.java
@@ -15,7 +15,7 @@
 package com.liferay.portal.upgrade.internal.index.updater.osgi.commands;
 
 import com.liferay.petra.string.StringBundler;
-import com.liferay.portal.util.BundleUtil;
+import com.liferay.portal.module.util.BundleUtil;
 import com.liferay.portal.util.IndexUpdaterUtil;
 
 import org.apache.felix.service.command.Descriptor;
